+++ pineappl/src/pids.rs
@@ -134,6 +134,19 @@ pub fn charge_conjugate_pdg_pid(pid: i32) -> i32 {
     }
 }
 
+/// Return the charge-conjugated particle ID of `pid` for the basis `lumi_id_types`. The returned
+/// tuple contains a factor that possible arises during the carge conjugation.
+#[must_use]
+pub fn charge_conjugate(lumi_id_types: &str, pid: i32) -> (i32, f64) {
+    match (lumi_id_types, pid) {
+        ("pdg_mc_ids", _) => (charge_conjugate_pdg_pid(pid), 1.0),
+        ("evol", 100 | 103 | 108 | 115 | 124 | 135) => (pid, 1.0),
+        ("evol", 200 | 203 | 208 | 215 | 224 | 235) => (pid, -1.0),
+        ("evol", _) => (charge_conjugate_pdg_pid(pid), 1.0),
+        _ => todo!(),
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;